fix(oidc-provider): authorize post-auth flow (#2849)
* oidcProvider authorize post-auth flow fix

Issue:

Recommended way to handle login for OIDC Provider flow is to call
authClient.signIn -- However, many of the callbacks in authorize.ts
endpoint directly throw their redirects, in many of these cases this
is not desired as it will redirect the fetch request when the intent
is to redirect the browser instead.

Fix:

If the call is coming from a cors (non-browser navigate) fetch call
when doing the consent page redirect or the final authorization
callback redirect, redirect using a hint in json payload to authClient
rather than attempting to redirect the CORS call.

This resolves issues both with CORS (that people were seeing) as
well as well as removing the need for the work around hack provided
in #2461

Tested: Locally, using updated code, oidcProvider and a genericOauth
client. Verified happy paths for login with / without scope, there
are some other issues with consent flow in general, but I'll address
those in a different PR.

* cleanup and lint

---------

Co-authored-by: Bereket Engida <bekacru@gmail.com>