fix: issue when setting rate limit storage with a different model name (#3536)
The `set` function of `createDbStorage` is referencing a hardcoded string 'rateLimit'. This causes an error when a custom name for the model has been used:

```
TypeError: Cannot read properties of undefined (reading 'fields')
``` 

It should use the model the `get` method uses which reads `modelName` if it has been specified.

This change fixes the error locally for me.